+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Datasets: delete
+
 ## [3.9.0] - 2016-09-21
 ### Added
 * Datasets API